+++ packages/common/package.json
@@ -28,7 +28,6 @@
   },
   "dependencies": {
     "@ceramicnetwork/docid": "^0.5.0",
-    "@ceramicnetwork/logger": "^0.1.0",
     "cids": "~1.1.6",
     "events": "^3.2.0",
     "lodash.clonedeep": "^4.5.0",
