+++ packages/common/src/index.ts
@@ -6,6 +6,7 @@ export * from './utils/doctype-utils'
 export * from './utils/test-utils'
 export * from './logger-provider'
 export * from './logger-provider-old' // TODO remove this
+export * from './loggers'
 export * from './pinning'
 export * from './doc-cache'
 export * from './unreachable-case-error'
