+++ packages/ipfs-daemon/package.json
@@ -30,7 +30,7 @@
   },
   "dependencies": {
     "@ceramicnetwork/ipfs-topology": "^0.0.2",
-    "@ceramicnetwork/logger": "^0.1.0",
+    "@ceramicnetwork/common": "^0.17.0-rc.0",
     "aws-sdk": "^2.782.0",
     "dag-jose": "^0.3.0",
     "datastore-s3": "^3.0.0",
