+++ packages/ipfs-daemon/src/healthcheck-server.ts
@@ -4,7 +4,7 @@ import express from "express"
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import type { IPFSAPI as IpfsApi } from 'ipfs-core/dist/src/components'
-import { DiagnosticsLogger } from '@ceramicnetwork/logger'
+import { DiagnosticsLogger } from '@ceramicnetwork/common'
 export type IpfsApi = typeof IpfsApi
 
 export class HealthcheckServer {
