+++ packages/ipfs-topology/package.json
@@ -24,7 +24,7 @@
     "clean": "rm -rf ./lib"
   },
   "dependencies": {
-    "@ceramicnetwork/logger": "^0.1.0",
+    "@ceramicnetwork/common": "^0.17.0-rc.0",
     "cross-fetch": "^3.0.6",
     "express": "^4.17.1"
   },
