+++ packages/logger/src/index.ts
@@ -1,2 +1,1 @@
-export * from "./loggers"
 export * from "./stream-helpers"
